requirements.txt
-e .
pytest
aws-cdk.core
aws-cdk.aws_iam
aws-cdk.aws_dynamodb
aws-cdk.aws_lambda
aws-cdk.aws_apigateway